Generate a valid typescript definition from the Unexpected instance.
Re-use logic from the instance builder and implement extra code to
process the returned set of assertions and their originals into TS
function definitions suitable for inclusion in a .d.ts template.

Basic support for extracting the types from all variants of all the
alternatives for each assertion is included. These tyes as found in
core are mapped to their TS equivalents and written as part of the
function definition output.