Generate the set of available assertions in base export at build time.
An issue with the previous version of this code was well as many of
previous "façade" implementations is that generating the conversion
glue resulted in highly dynamic code that made it rather difficult
to get a sense of what was implemented and how it was done.

Attempt to address this by ensuring a hard-coded list of supported
assertions is shipped as part of the built artefacts.