Address issues with no value assertions e.g. "to be truthy".
The function generation code assumed that all assertions would have
values and would thus pass them into the Unexpected assertion. That
is not true of all assertions though, so we need to account for this.

Port the type identification code from the typescript definition
builder and use it to generate an appropriate function when the
assertion has no value.