Rework .it. handling to ensure high quality output for errors.
The use of expect.it() as an implementation detail which backs
nested assertions would cause the anonymous bound it function to
be inspected when errors occurred. A user would not be able not
able to tell what was wrong, which was neither useful nor desired.

Make .it. functions returns instances of a small class that encode
the assertion to be performed. These instances record the assertion
string plus the arguments that would need to be passed to it.

Target the expect.it function with a custom type that identifies
the expect.it() wrapper functions and provides custom inspection
to remove them from the final output.