Rewrite handling of pass through "assertion" suffix with "to satisfy".
In Unexpected core there is a special case for "to satisfy" to allow
it to consume nested assertions by having a special version of the
assertion with an additional suffix. Given the way nested assertions
work with this camel cased API, we want to combine this into the main
assertion function.

Previosuly the way we dealt with this was to create a suffixed version
of all assertions that allow something nested on the RHS to, basically,
normalise this in the type processing code. Instead of that, alter the
assertion preparation code patch up the plain "to satisfy" so the type
determination doesn't need to worry about this.

As a side effect, we can complete rely on determined type information
to check whether nesting is allowed and as a conseuqnce a non-supported
variant is actually protected against.