Ensure nested assertions complete top-level assertions in afterEach().
Use a separate Unexpected instance to which we attach the afterEach()
hook consumed only by the top-level subject wrapping function. This
means the .it. functions are excluded from the counting mechanism
entirely. Finally arrange that the use of an expect.it() clears one
of the counted executing assertions thus making the test suite pass.