Protect against no assertion being performed on an subject.
Lay down infrastructure required to support an afterEach() hook.
Register this once at the point where the library is requred.

For each call to the top level expect function increment a counter
which effectively represents the number of assertions that should
occur. Arrange for the counter to be decremented on calls into any
of the assertion functions. In afterEach() check this pending count
and if it should be greater than zero then arrange an exception.