Fix ItPlaceholder occurring within an object spec using "to satisfy".
After 7b38595  .it. functions return a new ItPlaceholder() instance.
This was correctly unpacked by the top-level expect call but would not
be intepreted correctly when occurring within a "to satisfy" spec.

Adopt a trick form core - the wrapperObject type - defining a variant
of it for ItPlaceholder that causes it to be unwrapped and only then
be satisfied against the subject. The net effect is allowing the .it.
functions to be used within an object spec.

In order to maintain the valid JavaScript precedent for output extra
rendering logic is added on the type to output a named .it. function.