Implement a variant of declaration generation for Unexpected plugins.
Given a single named plugin we create two instances of Unexpected, one
with the plugin and the other without. We then compare these instances
and any assertions that are not common between them must be have been
added by the plugin.

We take this subset of plugin definitions and re-use the declaration
writing logic to serialise a plugin declaration using a special plugin
template.