Rework the tree and check in cased functions and declaration files.
This will allow us to preview the changes made by upgrading the
version of Unexpected (e.g. new assertions and types). It also
makes the construction of the library slightly less "magic" and
all executing code can be seen clearly in the tree. It is time.

Attempt to catch any changes that are not checked in by ensuring
bail out if we are unclean during prepublishOnly. This may need
to be revisited if we want to inline example as JSDoc comments.