Handle vararg assertion types when processing Unexpected instances.
Add detection for vararg types and, for now, to treat them as
a requirement for a singular argument of the determined type.
This avoid changes to TypeScript declaration generator (which
would likely need to output two function signatures) but makes
the assertions works correctly for now.

This commit also corrects invalid "to contain" declarations.